refactored executor code.
